+++ litebox_common_windows/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-int-enum = "1.2.0"
 object = { version = "0.36.7", default-features = false, features = ["pe", "read_core"] }
 thiserror = { version = "2.0.6", default-features = false }
 zerocopy = { version = "0.8", features = ["derive"] }
